Upgraded security

+++ laravel/laravel-giwu-api/app/Http/Controllers/PostController.php
@@ -3,10 +3,20 @@
 namespace App\Http\Controllers;
 
 use App\Models\Post;
+use Gate;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -25,7 +35,7 @@ class PostController extends Controller
             "body" => "required"
         ]);
 
-        $post = Post::create($field);
+        $post = $request->user()->posts()->create($field);
 
         return $post;
     }
@@ -43,6 +53,8 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        Gate::authorize("modify", $post);
+
         $field = $request->validate([
             "title" => "required|max:255",
             "body" => "required"
@@ -58,6 +70,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize("modify", $post);
+
         $post->delete();
 
         return ["message" => "The post is deleted"];
